+++ src/User/User.php
@@ -62,6 +62,8 @@ final class User
     private $legalIdentifier;
     /** @var string|null */
     private $loyaltyIdentifier;
+    /** @var string|null */
+    private $lang;
 
     /**
      * @internal
@@ -83,6 +85,7 @@ final class User
         $this->status = $data['status'] ?? null;
         $this->type = empty($data['type']) ? null : new UserType($data['type']);
         $this->phone = $data['phone'] ?? null;
+        $this->lang = $data['lang'] ?? null;
         $this->currencyCode = $data['currencyCode'] ?? null;
         $this->externalIdentifier = $data['externalIdentifier'] ?? '';
         $this->isProfessional = $data['isProfessional'] ?? null;
@@ -253,5 +256,10 @@ final class User
     public function getLoyaltyIdentifier(): ?string
     {
         return $this->loyaltyIdentifier;
+    }
+
+    public function getLanguage(): ?string
+    {
+        return $this->lang;
     }
 }